+++ packages/google_maps_flutter/google_maps_flutter/lib/google_maps_flutter.dart
@@ -18,6 +18,7 @@ export 'package:google_maps_flutter_platform_interface/google_maps_flutter_platf
         ArgumentCallbacks,
         ArgumentCallback,
         BitmapDescriptor,
+        BitmapScaling,
         CameraPosition,
         CameraPositionCallback,
         CameraTargetBounds,