+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/test/types/bitmap_test.dart
@@ -14,39 +14,118 @@ void main() {
     test('toJson / fromJson', () {
       final BitmapDescriptor descriptor =
           BitmapDescriptor.defaultMarkerWithHue(BitmapDescriptor.hueCyan);
-      final Object json = descriptor.toJson();
 
-      // Rehydrate a new bitmap descriptor...
-      // ignore: deprecated_member_use_from_same_package
-      final BitmapDescriptor descriptorFromJson =
-          BitmapDescriptor.fromJson(json);
-
-      expect(descriptorFromJson, isNot(descriptor)); // New instance
-      expect(identical(descriptorFromJson.toJson(), json), isTrue); // Same JSON
+      final Object expected = <Object>[
+        'defaultMarker',
+        BitmapDescriptor.hueCyan
+      ];
+      expect(descriptor.toJson(), equals(expected)); // Same JSON
     });
 
-    group('fromBytes constructor', () {
+    group('createFromAsset constructor', () {
+      test('without mipmaps', () async {
+        final BitmapDescriptor descriptor =
+            await BitmapDescriptor.createFromAsset(
+                ImageConfiguration.empty, 'path_to_asset_image',
+                mipmaps: false);
+        expect(descriptor, isA<BitmapDescriptor>());
+        expect(
+            descriptor.toJson(),
+            equals(<Object>[
+              'asset',
+              'path_to_asset_image',
+              BitmapDescriptor.bitmapAutoScaling,
+              1.0
+            ]));
+      });
+      test('with mipmaps', () async {
+        final BitmapDescriptor descriptor =
+            await BitmapDescriptor.createFromAsset(
+                ImageConfiguration.empty, 'path_to_asset_image',
+                // ignore: avoid_redundant_argument_values
+                mipmaps: true);
+        expect(descriptor, isA<BitmapDescriptor>());
+        expect(
+            descriptor.toJson(),
+            equals(<Object>[
+              'asset',
+              'path_to_asset_image',
+              BitmapDescriptor.bitmapAutoScaling,
+              1.0
+            ]));
+      });
+      test('with size and without mipmaps', () async {
+        final double devicePixelRatio = WidgetsBinding
+            .instance.platformDispatcher.views.first.devicePixelRatio;
+        const Size size = Size(100, 200);
+        final ImageConfiguration imageConfiguration =
+            ImageConfiguration(size: size, devicePixelRatio: devicePixelRatio);
+        final BitmapDescriptor descriptor =
+            await BitmapDescriptor.createFromAsset(
+                imageConfiguration, 'path_to_asset_image',
+                mipmaps: false);
+
+        expect(descriptor, isA<BitmapDescriptor>());
+        expect(
+            descriptor.toJson(),
+            equals(<Object>[
+              'asset',
+              'path_to_asset_image',
+              BitmapDescriptor.bitmapAutoScaling,
+              devicePixelRatio,
+              <double>[size.width, size.height]
+            ]));
+      });
+
+      test('with size and mipmaps', () async {
+        final double devicePixelRatio = WidgetsBinding
+            .instance.platformDispatcher.views.first.devicePixelRatio;
+        const Size size = Size(100, 200);
+        final ImageConfiguration imageConfiguration =
+            ImageConfiguration(size: size, devicePixelRatio: devicePixelRatio);
+        final BitmapDescriptor descriptor =
+            await BitmapDescriptor.createFromAsset(
+                imageConfiguration, 'path_to_asset_image',
+                // ignore: avoid_redundant_argument_values
+                mipmaps: true);
+
+        expect(descriptor, isA<BitmapDescriptor>());
+        expect(
+            descriptor.toJson(),
+            equals(<Object>[
+              'asset',
+              'path_to_asset_image',
+              BitmapDescriptor.bitmapAutoScaling,
+              1.0,
+              <double>[size.width, size.height]
+            ]));
+      });
+    });
+
+    group('createFromBytes constructor', () {
       test('with empty byte array, throws assertion error', () {
         expect(() {
-          BitmapDescriptor.fromBytes(Uint8List.fromList(<int>[]));
+          BitmapDescriptor.createFromBytes(Uint8List.fromList(<int>[]));
         }, throwsAssertionError);
       });
 
       test('with bytes', () {
-        final BitmapDescriptor descriptor = BitmapDescriptor.fromBytes(
+        final BitmapDescriptor descriptor = BitmapDescriptor.createFromBytes(
           Uint8List.fromList(<int>[1, 2, 3]),
         );
         expect(descriptor, isA<BitmapDescriptor>());
         expect(
             descriptor.toJson(),
             equals(<Object>[
-              'fromBytes',
+              'bytes',
               <int>[1, 2, 3],
+              BitmapDescriptor.bitmapAutoScaling,
+              1.0
             ]));
       });
 
-      test('with size, not on the web, size is ignored', () {
-        final BitmapDescriptor descriptor = BitmapDescriptor.fromBytes(
+      test('with size', () {
+        final BitmapDescriptor descriptor = BitmapDescriptor.createFromBytes(
           Uint8List.fromList(<int>[1, 2, 3]),
           size: const Size(40, 20),
         );
@@ -54,22 +133,10 @@ void main() {
         expect(
             descriptor.toJson(),
             equals(<Object>[
-              'fromBytes',
+              'bytes',
               <int>[1, 2, 3],
-            ]));
-      }, skip: kIsWeb);
-
-      test('with size, on the web, size is preserved', () {
-        final BitmapDescriptor descriptor = BitmapDescriptor.fromBytes(
-          Uint8List.fromList(<int>[1, 2, 3]),
-          size: const Size(40, 20),
-        );
-
-        expect(
-            descriptor.toJson(),
-            equals(<Object>[
-              'fromBytes',
-              <int>[1, 2, 3],
+              BitmapDescriptor.bitmapAutoScaling,
+              1.0,
               <int>[40, 20],
             ]));
       }, skip: !kIsWeb);