+++ packages/google_maps_flutter/google_maps_flutter_web/example/integration_test/google_maps_plugin_test.mocks.dart
@@ -278,13 +278,15 @@ class MockGoogleMapController extends _i1.Mock
         returnValueForMissingStub: null,
       );
   @override
-  void updateMarkers(_i2.MarkerUpdates? updates) => super.noSuchMethod(
+  _i3.Future<void> updateMarkers(_i2.MarkerUpdates? updates) =>
+      (super.noSuchMethod(
         Invocation.method(
           #updateMarkers,
           [updates],
         ),
-        returnValueForMissingStub: null,
-      );
+        returnValue: _i3.Future<void>.value(),
+        returnValueForMissingStub: _i3.Future<void>.value(),
+      ) as _i3.Future<void>);
   @override
   void showInfoWindow(_i2.MarkerId? markerId) => super.noSuchMethod(
         Invocation.method(