+++ packages/google_maps_flutter/google_maps_flutter_web/example/integration_test/markers_test.dart
@@ -38,7 +38,7 @@ void main() {
         const Marker(markerId: MarkerId('2')),
       };
 
-      controller.addMarkers(markers);
+      await controller.addMarkers(markers);
 
       expect(controller.markers.length, 2);
       expect(controller.markers, contains(const MarkerId('1')));
@@ -53,7 +53,7 @@ void main() {
       final Set<Marker> markers = <Marker>{
         const Marker(markerId: MarkerId('1')),
       };
-      controller.addMarkers(markers);
+      await controller.addMarkers(markers);
 
       marker = controller.markers[const MarkerId('1')]?.marker;
       expect(marker, isNotNull);
@@ -73,7 +73,8 @@ void main() {
           position: LatLng(42, 54),
         ),
       };
-      controller.changeMarkers(updatedMarkers);
+      await controller.changeMarkers(updatedMarkers);
+
       expect(controller.markers.length, 1);
 
       marker = controller.markers[const MarkerId('1')]?.marker;
@@ -136,7 +137,7 @@ void main() {
         const Marker(markerId: MarkerId('3')),
       };
 
-      controller.addMarkers(markers);
+      await controller.addMarkers(markers);
 
       expect(controller.markers.length, 3);
 
@@ -162,7 +163,7 @@ void main() {
         ),
       };
 
-      controller.addMarkers(markers);
+      await controller.addMarkers(markers);
 
       expect(controller.markers[const MarkerId('1')]?.infoWindowShown, isFalse);
 
@@ -188,7 +189,7 @@ void main() {
           infoWindow: InfoWindow(title: 'Title', snippet: 'Snippet'),
         ),
       };
-      controller.addMarkers(markers);
+      await controller.addMarkers(markers);
 
       expect(controller.markers[const MarkerId('1')]?.infoWindowShown, isFalse);
       expect(controller.markers[const MarkerId('2')]?.infoWindowShown, isFalse);
@@ -211,11 +212,11 @@ void main() {
       final Set<Marker> markers = <Marker>{
         Marker(
           markerId: const MarkerId('1'),
-          icon: BitmapDescriptor.fromBytes(bytes),
+          icon: BitmapDescriptor.createFromBytes(bytes),
         ),
       };
 
-      controller.addMarkers(markers);
+      await controller.addMarkers(markers);
 
       expect(controller.markers.length, 1);
       final gmaps.Icon? icon =
@@ -238,11 +239,12 @@ void main() {
       final Set<Marker> markers = <Marker>{
         Marker(
           markerId: const MarkerId('1'),
-          icon: BitmapDescriptor.fromBytes(bytes, size: const Size(20, 30)),
+          icon:
+              BitmapDescriptor.createFromBytes(bytes, size: const Size(20, 30)),
         ),
       };
 
-      controller.addMarkers(markers);
+      await controller.addMarkers(markers);
 
       expect(controller.markers.length, 1);
       final gmaps.Icon? icon =
@@ -271,7 +273,7 @@ void main() {
         ),
       };
 
-      controller.addMarkers(markers);
+      await controller.addMarkers(markers);
 
       expect(controller.markers.length, 1);
       final html.HtmlElement? content = controller.markers[const MarkerId('1')]
@@ -296,7 +298,7 @@ void main() {
         ),
       };
 
-      controller.addMarkers(markers);
+      await controller.addMarkers(markers);
 
       expect(controller.markers.length, 1);
       final html.HtmlElement? content = controller.markers[const MarkerId('1')]